+++ src/Generator.php
@@ -63,11 +63,15 @@ final class Generator implements Domain\Generator
         $now = time();
         $expiration = $expiresAt ? $expiresAt->getTimestamp() : $now + (60 * 60);
 
-        return $this->builder
+        $token = $this->builder
             ->setIssuedAt($now)
             ->setExpiration($expiration)
             ->sign($this->signer, $this->privateKey)
             ->getToken();
+
+        $this->builder->unsign();
+
+        return $token;
     }
 
     private function createBuilder(): Builder