+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+- Fix bug that not more than one custom token could be created at a time.
+
 ## 1.7.0 - 2018-01-03
 
 - Cache results from the HTTP Key Store in a PSR-16 cache (default: in memory)